[LTM-35] [ REFACTOR ] refactor Lotto detail view code into a separate LottoWinningInfoView file
