+++ LottoMate/LottoMate/Models/LottoResultModel.swift
@@ -22,7 +22,7 @@ struct LottoResultModel: Codable {
 
 /// 로또 추첨 결과 정보 모델 (당첨 번호, 당첨금 등)
 struct LottoResult: Codable {
-    /// 로또 회차
+    /// 로또 회차 키
     let lottoDrwNo: Int
     /// 복권 타입 (e.g. 'L645', 'L720', 'S500', 'S1000', 'S2000')
     let lottoType: String
