+++ LottoMate/LottoMate/Views/TabViews/WinningInfoDetail/WinningInfoDetailView+DrawView.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-extension WinningInfoDetailView {
+extension LottoWinningInfoView {
     func drawView() {
 //        lotteryDrawRound.text = ""
         styleLabel(for: lotteryDrawRound, fontStyle: .headline1, textColor: .primaryGray)
@@ -15,8 +15,8 @@ extension WinningInfoDetailView {
 //        drawDate.text = ""
         styleLabel(for: drawDate, fontStyle: .label2, textColor: .gray_ACACAC)
         
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(didTapDrawView))
-        lotteryDrawingInfo.addGestureRecognizer(tapGesture)
+//        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(didTapDrawView))
+//        lotteryDrawingInfo.addGestureRecognizer(tapGesture)
         lotteryDrawingInfo.isUserInteractionEnabled = true
         
         let previousRoundBtnImage = UIImage(named: "small_arrow_left")
