+++ LottoMate/LottoMate/Views/TabViews/WinningInfoDetail/WinningNumbersDetailViewController.swift
@@ -65,7 +65,7 @@ class WinningNumbersDetailViewController: UIViewController {
         view.addSubview(rootFlexContainer)
         
         // 회차별 로또 정보 가져오기 - 서버에 통신 줄이기 위해 주석처리
-//        viewModel.fetchLottoResult(round: 903)
+        viewModel.fetchLottoResult(round: 903)
         
         bindViewModel()
     }
@@ -77,7 +77,7 @@ class WinningNumbersDetailViewController: UIViewController {
     }
     
     private func bindViewModel() {
-        mainView.bind(viewModel: viewModel)
+//        mainView.bind(viewModel: viewModel)
         mainView.setupBindings(viewModel: viewModel)
     }
